cloudfx/6/requirements.txt
Flask==1.1.2
tweepy==3.8.0
pandas==1.0.3  
google-cloud-storage
google-cloud-bigquery
datetime
pandas_gbq
nltk
sklearn
scipy